+++ package.json
@@ -5,7 +5,7 @@
   "browser": "build/browser.js",
   "main": "build/node.js",
   "scripts": {
-    "test": "NODE_ENV=test node -r ts-node/register"
+    "test": "NODE_ENV=test node -r ts-node/register ./test/cli.ts"
   },
   "keywords": [],
   "author": "",
