--- src/happy-eyeballs.ts
@@ -1,3 +1,0 @@
-import { test } from "../test/test";
-import browser from './browser'
-import node from './node'
